Back

+++ back/src/main/java/US/SummerChallenge/NewsProject/Repository/NewsRepository.java
@@ -2,6 +2,7 @@ package US.SummerChallenge.NewsProject.Repository;
 
 import US.SummerChallenge.NewsProject.model.entity.News;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 /**
@@ -9,5 +10,5 @@ import org.springframework.stereotype.Repository;
  * @Date: 2021/6/29 0:58
  **/
 @Repository
-public interface NewsRepository extends JpaRepository<News,Integer> {
+public interface NewsRepository extends JpaRepository<News,Long> {
 }
